feat(volunteer-vacancies): Implement CRUD operations for managing volunteer vacancies and registrants
