+++ src/assets/icons/approve.jsx
@@ -1,53 +1,17 @@
-import React from "react";
-
-function ApproveIcon() {
+function ApproveIcon({ className }) {
   return (
     <svg
+      width="18"
+      height="17"
+      fill="currentColor"
+      viewBox="0 0 18 17"
+      className={className}
       xmlns="http://www.w3.org/2000/svg"
-      width="38"
-      height="39"
-      fill="none"
-      viewBox="0 0 38 39">
-      <g filter="url(#filter0_d_2691_4993)">
-        <rect
-          width="30"
-          height="30"
-          x="4"
-          y="0.5"
-          fill="#293066"
-          rx="4"
-          shapeRendering="crispEdges"></rect>
-        <path
-          fill="#fff"
-          d="M16.818 20.754a.346.346 0 00-.318-.203.341.341 0 00-.313.193l-.24.48a.5.5 0 01-.921-.066l-.526-1.577-.332.996a1.35 1.35 0 01-1.28.923H12.5a.5.5 0 110-1h.387a.35.35 0 00.333-.24l.568-1.707a.75.75 0 011.423 0l.434 1.301c.617-.506 1.69-.303 2.063.443a.355.355 0 00.292.195v-2.566l4-3.971V12.5h-4.25a.752.752 0 01-.75-.75V7.5h-6.25a.748.748 0 00-.75.75v14.5c0 .416.334.75.75.75h10.5c.416 0 .75-.334.75-.75V21.5l-4-.003a1.342 1.342 0 01-1.182-.743zM22 11.31a.748.748 0 00-.219-.528l-3.06-3.062a.75.75 0 00-.53-.219H18v4h4v-.19zm-3 7.034V20.5h2.156l5.052-5.087-2.12-2.121L19 18.342zm8.767-5.614l-.996-.996a.796.796 0 00-1.125 0l-.852.852 2.121 2.12.852-.85a.795.795 0 000-1.126z"></path>
-      </g>
-      <defs>
-        <filter
-          id="filter0_d_2691_4993"
-          width="38"
-          height="38"
-          x="0"
-          y="0.5"
-          colorInterpolationFilters="sRGB"
-          filterUnits="userSpaceOnUse">
-          <feFlood floodOpacity="0" result="BackgroundImageFix"></feFlood>
-          <feColorMatrix
-            in="SourceAlpha"
-            result="hardAlpha"
-            values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 127 0"></feColorMatrix>
-          <feOffset dy="4"></feOffset>
-          <feGaussianBlur stdDeviation="2"></feGaussianBlur>
-          <feComposite in2="hardAlpha" operator="out"></feComposite>
-          <feColorMatrix values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0"></feColorMatrix>
-          <feBlend
-            in2="BackgroundImageFix"
-            result="effect1_dropShadow_2691_4993"></feBlend>
-          <feBlend
-            in="SourceGraphic"
-            in2="effect1_dropShadow_2691_4993"
-            result="shape"></feBlend>
-        </filter>
-      </defs>
+    >
+      <path
+        fill="#fff"
+        d="M6.818 13.754a.346.346 0 00-.318-.203.34.34 0 00-.313.193l-.24.48a.5.5 0 01-.921-.066L4.5 12.581l-.332.996a1.35 1.35 0 01-1.28.923H2.5a.5.5 0 110-1h.387a.35.35 0 00.333-.24l.568-1.707a.75.75 0 011.424 0l.433 1.301c.617-.506 1.69-.303 2.063.443.059.118.171.186.292.195v-2.566l4-3.971V5.5H7.75A.752.752 0 017 4.75V.5H.75a.748.748 0 00-.75.75v14.5c0 .416.334.75.75.75h10.5c.416 0 .75-.334.75-.75V14.5l-4-.003a1.342 1.342 0 01-1.182-.743zM12 4.31a.748.748 0 00-.219-.528L8.721.72A.75.75 0 008.192.5H8v4h4v-.19zm-3 7.033V13.5h2.156l5.052-5.087-2.12-2.121L9 11.342zm8.767-5.613l-.996-.996a.796.796 0 00-1.125 0l-.852.852 2.121 2.12.852-.85a.795.795 0 000-1.126z"
+      ></path>
     </svg>
   );
 }
