+++ src/pages/list-registered/components/skeleton/skeleton-table.jsx
@@ -10,28 +10,30 @@ const SkeletonTable = () => {
         </TableCell>
         <TableCell>
           <div className="flex flex-col gap-2">
-            <Skeleton className="h-3 w-52" />
-            <Skeleton className="h-3 w-44" />
+            <Skeleton className="h-3 w-24" />
+            <Skeleton className="h-3 w-20" />
           </div>
         </TableCell>
         <TableCell>
           <div className="flex flex-col gap-2">
-            <Skeleton className="h-3 w-64" />
-            <Skeleton className="h-3 w-56" />
+            <Skeleton className="h-3 w-36" />
+            <Skeleton className="h-3 w-28" />
           </div>
         </TableCell>
         <TableCell>
-          <Skeleton className="h-5 w-5" />
+          <Skeleton className="h-3 w-28" />
+        </TableCell>
+        <TableCell>
+          <div className="flex flex-col gap-2">
+            <Skeleton className="h-3 w-36" />
+            <Skeleton className="h-3 w-28" />
+          </div>
         </TableCell>
         <TableCell>
           <Skeleton className="h-8 w-20 rounded-full" />
         </TableCell>
         <TableCell>
-          <div className="flex gap-2">
-            <Skeleton className="h-10 w-10" />
-            <Skeleton className="h-10 w-10" />
-            <Skeleton className="h-10 w-10" />
-          </div>
+          <Skeleton className="h-10 w-10" />
         </TableCell>
       </TableRow>
     </TableBody>
