--- src/pages/volunter/list-registered/response-form.jsx
@@ -1,21 +1,0 @@
-import FormLayout from "../components/form-layout";
-import Header from "@/components/header";
-import Layout from "@/components/layout";
-import React from "react";
-import FormHeader from "../components/form-header";
-import RegistrantForm from "./form-registrant";
-import { useParams } from "react-router-dom";
-
-function ResponseForm() {
-  return (
-    <Layout>
-      <Header titleHeader="Lowongan Relawan" />
-      <FormLayout>
-        <FormHeader />
-        <RegistrantForm />
-      </FormLayout>
-    </Layout>
-  );
-}
-
-export default ResponseForm;
